move smbios table creation out of propolis-standalone
