+++ lib/propolis/src/firmware/smbios/mod.rs
@@ -5,6 +5,7 @@
 use std::collections::BTreeMap;
 use std::fmt;
 
+use crate::cpuid;
 use table::{Table, Type127};
 
 mod bits;
@@ -196,3 +197,134 @@ impl TryFrom<String> for SmbString {
 #[derive(thiserror::Error, Debug)]
 #[error("String contains NUL byte")]
 pub struct SmbStringNulError();
+
+pub struct SmbiosParams {
+    pub memory_size: usize,
+    pub rom_size: usize,
+    pub rom_release_date: String,
+    pub rom_version: String,
+    pub num_cpus: u8,
+    pub cpuid_ident: Option<cpuid::Entry>,
+    pub cpuid_procname: Option<[cpuid::Entry; 3]>,
+}
+
+impl SmbiosParams {
+    pub fn generate_table(&self) -> anyhow::Result<TableBytes> {
+        use table::{type0, type1, type16, type4};
+        let bios_version = self
+            .rom_version
+            .as_str()
+            .try_into()
+            .expect("bootrom version string doesn't contain NUL bytes");
+        let smb_type0 = table::Type0 {
+            vendor: "Oxide".try_into().unwrap(),
+            bios_version,
+            bios_release_date: self.rom_release_date.as_str().try_into().unwrap(), // TODO: "Bureaucracy 41, 3186 YOLD".try_into().unwrap(),
+            bios_rom_size: ((self.rom_size / (64 * 1024)) - 1) as u8,
+            bios_characteristics: type0::BiosCharacteristics::UNSUPPORTED,
+            bios_ext_characteristics: type0::BiosExtCharacteristics::ACPI
+                | type0::BiosExtCharacteristics::UEFI
+                | type0::BiosExtCharacteristics::IS_VM,
+            ..Default::default()
+        };
+
+        let smb_type1 = table::Type1 {
+            manufacturer: "Oxide".try_into().unwrap(),
+            product_name: "OxVM".try_into().unwrap(),
+            wake_up_type: type1::WakeUpType::PowerSwitch,
+            ..Default::default()
+        };
+
+        let cpuid_vendor = cpuid::host_query(cpuid::Ident(0x0, None));
+        let cpuid_ident = self
+            .cpuid_ident
+            .unwrap_or_else(|| cpuid::host_query(cpuid::Ident(0x1, None)));
+        let family = match cpuid_ident.eax & 0xf00 {
+            // If family ID is 0xf, extended family is added to it
+            0xf00 => (cpuid_ident.eax >> 20 & 0xff) + 0xf,
+            // ... otherwise base family ID is used
+            base => base >> 8,
+        };
+
+        let vendor = cpuid::VendorKind::try_from(cpuid_vendor);
+        let proc_manufacturer = match vendor {
+            Ok(cpuid::VendorKind::Intel) => "Intel",
+            Ok(cpuid::VendorKind::Amd) => "Advanced Micro Devices, Inc.",
+            _ => "",
+        }
+        .try_into()
+        .unwrap();
+        let proc_family = match (vendor, family) {
+            // Zen
+            (Ok(cpuid::VendorKind::Amd), family) if family >= 0x17 => 0x6b,
+            //unknown
+            _ => 0x2,
+        };
+        let proc_id = u64::from(cpuid_ident.eax) | u64::from(cpuid_ident.edx) << 32;
+        let procname_entries = self.cpuid_procname.or_else(|| {
+            if cpuid::host_query(cpuid::Ident(0x8000_0000, None)).eax >= 0x8000_0004
+            {
+                Some([
+                    cpuid::host_query(cpuid::Ident(0x8000_0002, None)),
+                    cpuid::host_query(cpuid::Ident(0x8000_0003, None)),
+                    cpuid::host_query(cpuid::Ident(0x8000_0004, None)),
+                ])
+            } else {
+                None
+            }
+        });
+        let proc_version = procname_entries
+            .and_then(|e| cpuid::parse_brand_string(e).ok())
+            .unwrap_or("".to_string());
+
+        let smb_type4 = table::Type4 {
+            proc_type: type4::ProcType::Central,
+            proc_family,
+            proc_manufacturer,
+            proc_id,
+            proc_version: proc_version.as_str().try_into().unwrap_or_default(),
+            status: type4::ProcStatus::Enabled,
+            // unknown
+            proc_upgrade: 0x2,
+            // make core and thread counts equal for now
+            core_count: self.num_cpus,
+            core_enabled: self.num_cpus,
+            thread_count: self.num_cpus,
+            proc_characteristics: type4::Characteristics::IS_64_BIT
+                | type4::Characteristics::MULTI_CORE,
+            ..Default::default()
+        };
+
+        let mut smb_type16 = table::Type16 {
+            location: type16::Location::SystemBoard,
+            array_use: type16::ArrayUse::System,
+            error_correction: type16::ErrorCorrection::Unknown,
+            num_mem_devices: 1,
+            ..Default::default()
+        };
+        smb_type16.set_max_capacity(self.memory_size);
+        let phys_mem_array_handle = 0x1600.into();
+
+        let mut smb_type17 = table::Type17 {
+            phys_mem_array_handle,
+            // Unknown
+            form_factor: 0x2,
+            // Unknown
+            memory_type: 0x2,
+            ..Default::default()
+        };
+        smb_type17.set_size(Some(self.memory_size));
+
+        let smb_type32 = table::Type32::default();
+
+        let mut smb_tables = Tables::new(0x7f00.into());
+        smb_tables.add(0x0000.into(), &smb_type0).unwrap();
+        smb_tables.add(0x0100.into(), &smb_type1).unwrap();
+        smb_tables.add(0x0300.into(), &smb_type4).unwrap();
+        smb_tables.add(phys_mem_array_handle, &smb_type16).unwrap();
+        smb_tables.add(0x1700.into(), &smb_type17).unwrap();
+        smb_tables.add(0x3200.into(), &smb_type32).unwrap();
+
+        Ok(smb_tables.commit())
+    }
+}
